# services/main/docs/swagger.yaml
definitions:
...
  handlers.CreatePaymentAccountRequest:
    properties:
      identifier:
        example: "0241234567"
        type: string
      is_default:
        example: false
        type: boolean
      metadata:
        additionalProperties: {}
        type: object
      provider:
        enum:
        - MTN
        - VODAFONE
        - AIRTELTIGO
        example: MTN
        type: string
      rail:
        enum:
        - MOMO
        - BANK_TRANSFER
        - CARD
        - OFFLINE
        example: MOMO
        type: string
      status:
        enum:
        - ACTIVE
        - DISABLED
        example: ACTIVE
        type: string
    required:
    - rail
    - status
    type: object
...
  handlers.UpdatePaymentAccountRequest:
    properties:
      identifier:
        example: "0241234567"
        type: string
      is_default:
        example: true
        type: boolean
      metadata:
        additionalProperties: {}
        type: object
      provider:
        example: MTN
        type: string
      status:
        enum:
        - ACTIVE
        - DISABLED
        example: ACTIVE
        type: string
    type: object
...
  lib.HTTPError:
    properties:
      errors:
        properties:
          message:
            type: string
        type: object
    type: object
  lib.HTTPReturnPaginatedMetaResponse:
    properties:
      has_next_page:
        example: true
        type: boolean
      has_previous_page:
        example: false
        type: boolean
      order:
        example: desc
        type: string
      order_by:
        example: created_at
        type: string
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  transformations.OutputAdmin:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: admin@example.com
        type: string
      id:
        example: 4fce5dc8-8114-4ab2-a94b-b4536c27f43b
        type: string
      name:
        example: Admin Name
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  transformations.OutputClient:
    properties:
      address:
        example: 123 Main St, Suite 100
        type: string
      city:
        example: San Francisco
        type: string
      client_application:
        $ref: '#/definitions/transformations.OutputClientApplication'
      client_application_id:
        example: app-1234
        type: string
      country:
        example: US
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      id:
        example: 4fce5dc8-8114-4ab2-a94b-b4536c27f43b
        type: string
      latitude:
        example: 37.7749
        type: number
      longitude:
        example: -122.4194
        type: number
      name:
        example: Acme Corp
        type: string
      region:
        example: California
        type: string
      sub_type:
        example: LANDLORD
        type: string
      support_email:
        example: support@somewebiste.com
        type: string
      support_phone:
        example: "+233551235555"
        type: string
      type:
        example: INDIVIDUAL
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      website_url:
        example: https://www.somewebiste.com
        type: string
    type: object
  transformations.OutputClientApplication:
    properties:
      address:
        description: company address or individual home address
        example: 21st Neon Street
        type: string
      approved_by_id:
        example: S90092
        type: string
      city:
        example: Accra
        type: string
      contact_email:
        example: contact@email.com
        type: string
      contact_name:
        example: John Doe
        type: string
      contact_phone_number:
        example: "01234567890"
        type: string
      country:
        example: Ghana
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      date_of_birth:
        description: individual specific fields
        example: "2025-01-31"
        type: string
      description:
        example: Taking you to the next level!
        type: string
      id:
        example: 4fce5dc8-8114-4ab2-a94b-b4536c27f43b
        type: string
      id_document_url:
        example: www.id-doc-url.com/id.pdf
        type: string
      id_expiry:
        example: "2040-12-31"
        type: string
      id_number:
        example: GHA-123-456-7890
        type: string
      id_type:
        example: GHANACARD
        type: string
      latitude:
        example: 5.6037
        type: number
      logo_url:
        example: www.logo-url.com/logo.png
        type: string
      longitude:
        example: -0.187
        type: number
      name:
        example: Company Name
        type: string
      region:
        example: Greater Accra
        type: string
      registration_number:
        description: company specific fields
        example: GR1234567890
        type: string
      rejected_because:
        example: No reason
        type: string
      rejected_by:
        $ref: '#/definitions/transformations.OutputAdmin'
      rejected_by_id:
        example: R234110
        type: string
      status:
        example: ClientApplication.Status.Approved
        type: string
      sub_type:
        example: ESTATE MANAGER
        type: string
      support_email:
        example: support@email.com
        type: string
      support_phone:
        example: +233 (0)12 345 6789
        type: string
      type:
        example: COMPANY
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      website_url:
        example: www.company-url.com
        type: string
    type: object
...
  transformations.OutputPaymentAccount:
    properties:
      client:
        $ref: '#/definitions/transformations.OutputClient'
      client_id:
        example: b50874ee-1a70-436e-ba24-572078895982
        type: string
      created_at:
        example: "2023-01-01T00:00:00Z"
        format: date-time
        type: string
      id:
        example: 4fce5dc8-8114-4ab2-a94b-b4536c27f43b
        format: uuid
        type: string
      identifier:
        example: "0241234567"
        type: string
      is_default:
        example: false
        type: boolean
      metadata:
        additionalProperties: {}
        type: object
      owner_type:
        example: PROPERTY_OWNER
        type: string
      provider:
        example: MTN
        type: string
      rail:
        example: MOMO
        type: string
      status:
        example: ACTIVE
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        format: date-time
        type: string
    type: object
...
info:
  contact:
    email: domeybenjamin1@gmail.com
    name: Domey Benjamin
    url: https://www.linkedin.com/in/benjamin-armah-kesey-domey/
  description: This is the RentLoop API documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://rentloop.com/terms
  title: RentLoop API
  version: "1.0"
paths:
...
  /api/v1/admin/payment-accounts:
    get:
      consumes:
      - application/json
      description: Get all payment accounts for the current client (Admin)
      parameters:
      - in: query
        name: end_date
        type: string
      - collectionFormat: multi
        example:
        - a8098c1a-f86e-11da-bd1a-00112444be1e
        in: query
        items:
          type: string
        name: ids
        type: array
      - in: query
        name: is_default
        type: boolean
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - in: query
        name: order_by
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: owner_types
        type: array
      - in: query
        minimum: 0
        name: page
        type: integer
      - in: query
        minimum: 0
        name: page_size
        type: integer
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: populate
        type: array
      - enum:
        - MTN
        - VODAFONE
        - AIRTELTIGO
        - BANK_API
        in: query
        name: provider
        type: string
      - in: query
        name: query
        type: string
      - enum:
        - MOMO
        - BANK_TRANSFER
        - CARD
        - OFFLINE
        in: query
        name: rail
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        minItems: 1
        name: search_fields
        type: array
      - in: query
        name: start_date
        type: string
      - enum:
        - ACTIVE
        - DISABLED
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  meta:
                    $ref: '#/definitions/lib.HTTPReturnPaginatedMetaResponse'
                  rows:
                    items:
                      $ref: '#/definitions/transformations.OutputPaymentAccount'
                    type: array
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all payment accounts (Admin)
      tags:
      - Payment Accounts
    post:
      consumes:
      - application/json
      description: Create a new payment account for a client (Admin)
      parameters:
      - description: Create Payment Account Request Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreatePaymentAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Payment account created successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/transformations.OutputPaymentAccount'
            type: object
        "400":
          description: Error occurred when creating a payment account
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Invalid or absent authentication token
          schema:
            type: string
        "500":
          description: An unexpected error occurred
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Creates a new payment account (Admin)
      tags:
      - Payment Accounts
  /api/v1/admin/payment-accounts/{payment_account_id}:
    delete:
      consumes:
      - application/json
      description: Delete a payment account (Admin)
      parameters:
      - description: Payment Account ID
        in: path
        name: payment_account_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Payment account deleted successfully
        "400":
          description: Error occurred when deleting a payment account
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Invalid or absent authentication token
          schema:
            type: string
        "500":
          description: An unexpected error occurred
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete a payment account (Admin)
      tags:
      - Payment Accounts
    get:
      consumes:
      - application/json
      description: Get payment account by ID (Admin)
      parameters:
      - description: Payment Account ID
        in: path
        name: payment_account_id
        required: true
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: populate
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/transformations.OutputPaymentAccount'
            type: object
        "400":
          description: Error occurred when fetching a payment account
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Invalid or absent authentication token
          schema:
            type: string
        "404":
          description: Payment account not found
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: An unexpected error occurred
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get payment account by ID (Admin)
      tags:
      - Payment Accounts
    patch:
      consumes:
      - application/json
      description: Update an existing payment account (Admin)
      parameters:
      - description: Payment Account ID
        format: uuid4
        in: path
        name: payment_account_id
        required: true
        type: string
      - description: Payment account details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdatePaymentAccountRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Payment account updated successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/transformations.OutputPaymentAccount'
            type: object
        "400":
          description: Error occurred when updating a payment account
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Invalid or absent authentication token
          schema:
            type: string
        "404":
          description: Payment account not found
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: An unexpected error occurred
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update an existing payment account (Admin)
      tags:
      - Payment Accounts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"